backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
ffmpeg-python
openai-whisper
opencv-python
numpy
pillow
python-dotenv
httpx
python-multipart
